backend/requirements.txt
# backend/requirements.txt

fastapi
numpy
opencv-python
pillow
python-multipart
uvicorn
onnxruntime
# onnxruntime-gpu  # if have a NVIDIA GPU
aiofiles